requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
alembic==1.14.0
sqlalchemy==2.0.36
pyyaml>=6.0
requests>=2.32
pytest>=8.0
pytest-json-report>=1.5
schemathesis>=3.37
playwright>=1.44
